fix(tests): simplify fixtures for dotnet test compat
