+++ Brotli.NET/Brotli.NET.Tests/BasicCompressionTest.cs
