+++ Brotli.NET/Brotli.NET.Tests/Resource/BingCN.bin
