+++ Brotli.NET/Brotli.NET.Tests/Resource/BingCN_Compressed.bin
