+++ Brotli.NET/Brotli.NET.sln
@@ -1,13 +1,13 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
-VisualStudioVersion = 17.13.35825.156 d17.13
+VisualStudioVersion = 17.13.35825.156
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Brotli.NET.Tests", "TestBrotli\Brotli.NET.Tests.csproj", "{1BB4F249-0ED5-49E8-97EC-4A5FF495B296}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Brotli.CompressionProvider.Sample", "Brotli.CompressionProvider.Sample\Brotli.CompressionProvider.Sample.csproj", "{5AD5F247-6913-47E7-8F46-1159E0843FC8}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Brotli.Core", "Brotli.Core\Brotli.Core.csproj", "{8752A96F-E0D9-4185-8A53-CA9C98679549}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Brotli.NET.Tests", "Brotli.NET.Tests\Brotli.NET.Tests.csproj", "{1A4D376F-883B-D69E-E4E2-399AEA647972}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,10 +15,6 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{1BB4F249-0ED5-49E8-97EC-4A5FF495B296}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1BB4F249-0ED5-49E8-97EC-4A5FF495B296}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1BB4F249-0ED5-49E8-97EC-4A5FF495B296}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1BB4F249-0ED5-49E8-97EC-4A5FF495B296}.Release|Any CPU.Build.0 = Release|Any CPU
 		{5AD5F247-6913-47E7-8F46-1159E0843FC8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{5AD5F247-6913-47E7-8F46-1159E0843FC8}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{5AD5F247-6913-47E7-8F46-1159E0843FC8}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -27,6 +23,10 @@ Global
 		{8752A96F-E0D9-4185-8A53-CA9C98679549}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{8752A96F-E0D9-4185-8A53-CA9C98679549}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{8752A96F-E0D9-4185-8A53-CA9C98679549}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1A4D376F-883B-D69E-E4E2-399AEA647972}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1A4D376F-883B-D69E-E4E2-399AEA647972}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1A4D376F-883B-D69E-E4E2-399AEA647972}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1A4D376F-883B-D69E-E4E2-399AEA647972}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
